pyproject.toml
[project]
name = "wall-mounted-backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "django==5.1.6",
    "djangorestframework==3.15.2",
    "pillow==11.1.0",
    "psycopg==3.2.4",
    "python-dotenv==1.0.1",
]
